+++ shoop/admin/modules/products/views/__init__.py
@@ -6,6 +6,7 @@
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
 
+from .archive import ProductArchiveView
 from .delete import ProductDeleteView
 from .edit import ProductEditView
 from .edit_cross_sell import ProductCrossSellEditView
@@ -15,6 +16,7 @@ from .edit_variation import ProductVariationView
 from .list import ProductListView
 
 __all__ = [
+    "ProductArchiveView",
     "ProductCrossSellEditView",
     "ProductDeleteView",
     "ProductEditView",